requirements.txt
streamlit==1.44.1
pandas==2.2.3
scikit-learn==1.6.1
nltk==3.9.1
joblib==1.4.2
python-dotenv